+++ src/main/java/org/roda/rodain/rules/Rule.java
@@ -6,10 +6,11 @@ import javafx.scene.image.Image;
 import javafx.scene.paint.Color;
 import org.roda.rodain.core.PathCollection;
 import org.roda.rodain.rules.filters.ContentFilter;
-import org.roda.rodain.rules.sip.*;
+import org.roda.rodain.sip.*;
 import org.roda.rodain.schema.DescriptionObject;
 import org.roda.rodain.schema.ui.SchemaNode;
 import org.roda.rodain.schema.ui.SipPreviewNode;
+import org.roda.rodain.sip.creators.*;
 import org.roda.rodain.source.ui.items.SourceTreeDirectory;
 import org.roda.rodain.source.ui.items.SourceTreeItem;
 import org.roda.rodain.source.ui.items.SourceTreeItemState;
