+++ src/main/java/org/roda/rodain/rules/TemplateToForm.java
@@ -2,7 +2,7 @@ package org.roda.rodain.rules;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Template;
-import org.roda.rodain.rules.sip.MetadataValue;
+import org.roda.rodain.sip.MetadataValue;
 
 import java.io.IOException;
 import java.util.HashMap;
