+++ src/main/java/org/roda/rodain/rules/ui/RuleModalController.java
@@ -8,8 +8,8 @@ import org.roda.rodain.rules.MetadataOptions;
 import org.roda.rodain.rules.Rule;
 import org.roda.rodain.rules.RuleTypes;
 import org.roda.rodain.rules.VisitorStack;
-import org.roda.rodain.rules.sip.SipPreview;
-import org.roda.rodain.rules.sip.SipPreviewCreator;
+import org.roda.rodain.sip.SipPreview;
+import org.roda.rodain.sip.creators.SipPreviewCreator;
 import org.roda.rodain.schema.ui.SchemaNode;
 import org.roda.rodain.source.ui.items.SourceTreeItem;
 import org.roda.rodain.utils.ModalStage;
