+++ src/main/java/org/roda/rodain/sip/MetadataValue.java
@@ -1,4 +1,4 @@
-package org.roda.rodain.rules.sip;
+package org.roda.rodain.sip;
 
 import org.roda.rodain.core.I18n;
