+++ src/main/java/org/roda/rodain/sip/PseudoSIP.java
@@ -1,4 +1,4 @@
-package org.roda.rodain.rules.sip;
+package org.roda.rodain.sip;
 
 import org.roda.rodain.rules.TreeNode;
