+++ src/main/java/org/roda/rodain/sip/creators/SipPreviewCreator.java
@@ -1,10 +1,12 @@
-package org.roda.rodain.rules.sip;
+package org.roda.rodain.sip.creators;
 
 import org.apache.commons.io.FilenameUtils;
 import org.roda.rodain.rules.MetadataOptions;
 import org.roda.rodain.rules.TreeNode;
 import org.roda.rodain.rules.filters.ContentFilter;
 import org.roda.rodain.schema.DescObjMetadata;
+import org.roda.rodain.sip.SipPreview;
+import org.roda.rodain.sip.SipRepresentation;
 import org.roda.rodain.utils.TreeVisitor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
